requirements.txt
pandas
numpy
scikit-learn
tableshift
scipy
oumi-sdk